#192 #574 make openssl optional
